requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

streamlit==1.31.0

pillow==10.2.0
pdf2image==1.17.0
pytesseract==0.3.10
easyocr==1.7.1
opencv-python-headless==4.9.0.80

transformers==4.36.2
torch==2.2.1
#sentencepiece==0.1.96

python-dotenv==1.0.1
requests==2.31.0
huggingface-hub==0.20.3

langchain
langchain_groq
python-dotenv
transformers
torch
PyMuPDF
